requirements.txt
absl-py
alabaster==0.7.12
alembic==1.5.6
amqp
#apex @ file:///home/vamsi/Academics/BITS/UCL/SSG/src/ProtBert/apex
appdirs==1.4.4
#apturl==0.5.2
arrow==0.8.0
ase==3.19.1
asgiref==3.2.7
astunparse
attrs==19.3.0
Babel==2.9.0
backcall==0.2.0
beniget~=0.4.0
billiard<5.0,>=4.2.0
bio
#bio-embeddings
bio-embeddings-allennlp
bio-embeddings-bepler==0.0.1
#bio-embeddings-CPCProt
bio-embeddings-esm==0.2.1
#bio-embeddings-plus
bio-embeddings-tape-proteins==0.5
biopython<2.0,>=1.79
#biovec
bleach==3.1.5
blinker==1.4
blis==0.7.4
bokeh==2.1.1
boto==2.46.1
boto3==1.14.48
botocore==1.17.48
#Brlapi==0.7.0
bz2file==0.98
cachetools==4.1.0
catalogue==1.0.0
cathpy==0.3.10
celery<6.0.0,>=5.2.7
certifi==2020.4.5.2
chardet==3.0.4
click>=8.0
click-didyoumean>=0.3.0
click-plugins==1.1.1
click-repl>=0.2.0
cliff==3.7.0
cloudpickle==1.5.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==4.7.2
#command-not-found==0.3
ConfigArgParse==0.11.0
conllu==4.4
cryptography==2.8
#cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
Cython
dask==2.21.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.4
defusedxml==0.6.0
DendroPy==4.5.1
#dgl-cu101==0.6.1
dgllife==0.2.8
distlib==0.3.1
distributed==2.21.0
distro==1.4.0
#distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-heroku==0.3.1
docutils==0.15.2
EasyProcess==0.3
editdistance==0.5.3
empath==0.89
#Endgame-Singularity==1.0b1
entrypoints==0.3
et-xmlfile==1.0.1
filelock
flaky==3.7.0
Flask==2.1.3
Flask-Cors<4.0.0,>=3.0.10
flatbuffers>=23.5.26
fsspec>=2023.5.0
ftfy==5.9
future==0.18.2
gast~=0.6.0
#gensim==4.3.2
gevent
google-auth>=2.15.0
google-auth-oauthlib<1.1,>=0.5
google-pasta>=0.1.1
googledrivedownloader==0.4
greenlet>=2.0.0
grpcio>=1.24.3,<2.0
gym==0.26.2
h5py<4.0.0,>=3.7.0 
HeapDict==1.0.1
httplib2==0.14.0
hyperopt==0.2.4
idna==2.7
imagesize==1.2.0
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata<5.0.0,>=4.11.4
iniconfig==1.0.1
intervaltree==3.1.0
ipykernel==5.3.0
ipympl==0.5.6
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate<1.0.0,>=0.7.2
itsdangerous>=2.0
jax==0.1.77
jax-unirep==1.0.3
jaxlib>=0.1.57,<0.4.7
jdcal==1.4.1
jedi==0.17.0
Jinja2
jmespath==0.10.0
joblib>=1.2.0
jsonnet==0.17.0
jsonpickle==1.5.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
keras<2.15,>=2.14.0
Keras-Applications==1.0.8
keras-preprocessing>=1.1.1
keras-self-attention==0.49.0
keyring==18.0.1
kiwisolver>=1.3.1
# kombu<6.0,>=5.3.4
kombu==5.3.7
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazypredict==0.2.7
lightgbm==2.3.1
llvmlite
lmdb
lock==2018.3.25.2110
locket==0.2.0
Logbook==1.5.3
#louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.4
Markdown==3.2.2
MarkupSafe>=2.0
matplotlib>=3.5.2
mistune==0.8.4
more-itertools
mrcfile==1.1.2
msgpack==1.0.0
multipledispatch==0.6.0
munkres==1.1.2
murmurhash==1.0.5
nbconvert==5.6.1
nbformat==5.0.7
netifaces
networkx==2.3
nltk
notebook==6.0.3
numba
numexpr
numpy<2.0.0,>=1.19.2
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
opencv-python
openpyxl==3.0.3
opt-einsum>=2.3.2
optuna
overrides<5,>=4
packaging
pandas<2.0.0,>=1.0.3
pandocfilters
parsimonious==0.8.1
parso==0.7.0
partd==1.1.0
pbr==5.5.1
pdf2image==1.14.0
pexpect==4.6.0
pickleshare==0.7.5
pillow>=8
plac==1.1.3
plotly>=5.8.0
pluggy<2.0,>=1.5
ply==3.11
plyfile==0.7.2
polib==1.1.0
preshed==3.0.5
prettytable==2.1.0
prometheus-client==0.8.0
prompt-toolkit>=3.0.36
PROPKA==3.1.0
protobuf<4.0.0,>=3.19.6
psutil==5.7.2
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups
pydicom==2.0.0
pyfasta==0.5.2
pygame
pythran
#pyGaussDCA
pyglet==1.5.0
pygments<3.0.0,>=2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.11.2
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.16.0
pytesseract==0.3.6
pytest
#python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil
#python-debian===0.1.36ubuntu1
python-editor==1.0.4
pytorch-ignite==0.4.4
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz
PyVirtualDisplay==1.3.2
pyxdg==0.26
PyYAML==6.0
pyzmq
qtconsole==4.7.4
QtPy==1.9.0
rdflib
rdkit-pypi
redis==3.5.3
regex>=2021.8.3
reportlab==3.5.34
requests
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
responses==0.12.1
retrying==1.3.3
rsa==4.0
ruamel.yaml<0.18.0,>=0.17.10
ruamel.yaml.clib>=0.2.7
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn<2.0.0,>=1.2.2
scipy>=1.8.1
#screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash
sentencepiece==0.1.96
simplejson==3.16.0
six>=1.12.0
sklearn==0.0
smart-open
snowballstemmer==2.1.0
sortedcontainers==2.2.2
spacy==2.3.5
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.23
sqlparse==0.3.1
srsly==1.0.5
ssh-import-id==5.10
stevedore==3.3.0
swifter==0.305
systemd-python==234
tables==3.6.1
tb-nightly==2.18.0a20240616
tblib==1.6.0
tensorboard<2.15,>=2.14
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.14.0
tensorflow-addons
tensorflow-estimator<2.15,>=2.14.0
#tensorflow-gpu==2.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.8.0.dev2021122109
tflearn==0.3.1
thinc==7.4.5
threadpoolctl>=3.1.0
tokenizers<0.21,>=0.20
toml==0.10.1
toolz==0.10.0
torch>=1.8.0,<=1.9.1
#torch-cluster==1.6.3
torch-geometric==1.5.0
#torch-scatter==2.1.2
#torch-sparse==0.6.18
#torch-spline-conv==1.2.2
torchvision==0.10.1
tornado
tqdm<5.0.0,>=4.64.0
traitlets
transformers
typeguard==2.9.1
typing-extensions>=3.7.4.3
#ubuntu-advantage-tools==20.3
#ubuntu-drivers-common==0.0.0
#ufw==0.36
umap-learn>=0.5.3
#unattended-upgrades==0.1
Unidecode==1.2.0
urllib3==1.25.10
#usb-creator==0.3.7
vine<6.0,>=5.1.0
virtualenv==20.0.27
wadllib==1.3.3
wasabi==0.8.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==2.0.3
wget==3.2
whitenoise==5.1.0
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.12.1
xgboost==2.1.1
#xkit==0.0.0
xlrd==1.2.0
yellowbrick
zict==2.0.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.2.0
ml-dtypes==0.2.0
